Fold single-function modules into the shared generator utilities
Two modules each held one function with a single caller:

- clientele/utils.py only contained get_client_project_directory_path,
  which is used exclusively by the generators at code-generation time.
  It now lives in clientele/generators/shared/utils.py with the rest of
  the generator helpers, so the top-level package no longer has a
  utils.py whose contents are unrelated to the runtime library.
- clientele/generators/schema_utils.py only contained
  build_union_type_string, called once from the shared SchemasGenerator.
  It is now defined in clientele/generators/shared/schemas.py next to
  its caller.
